Module massa_versioning_worker::versioning

Structs

Active

State machine for a Versioning component that tracks the deployment state

Advance

A message to update the ComponentState

Defined

State machine for a Versioning component that tracks the deployment state

Failed

State machine for a Versioning component that tracks the deployment state

LockedIn

State machine for a Versioning component that tracks the deployment state

MipInfo

MIP info (name & versions & time range for a MIP)

MipState

Wrapper of ComponentState (in order to keep state history)

MipStatsConfig

Statistics in MipStoreRaw

MipStore

Database for all MIP info

MipStoreRaw

Store of all versioning info

MipStoreStats 🔒

In order for a MIP to be accepted, we compute statistics about other node ‘network’ version announcement

Started

State machine for a Versioning component that tracks the deployment state

Enums

ComponentState

State machine for a Versioning component that tracks the deployment state

ComponentStateMessages

ComponentStateTypeId

MipComponent

Versioning component enum

StateAtError

Error returned by MipStateHistory::state_at

UpdateWithError

Error returned by

Constants

Active__num_enum_0__ 🔒

Address__num_enum_0__ 🔒

Block__num_enum_0__ 🔒

Defined__num_enum_0__ 🔒

Error__num_enum_0__ 🔒

Failed__num_enum_0__ 🔒

LockedIn__num_enum_0__ 🔒

Started__num_enum_0__ 🔒

VM__num_enum_0__ 🔒

__Nonexhaustive__num_enum_0__ 🔒

Statics

CALLSITE 🔒

META 🔒